third_party/blink/web_tests/platform/mac/external/wpt/css/css-grid/masonry/tentative/track-sizing/auto-repeat/row-auto-repeat-009-expected.txt

This is a testharness.js-based test.
[FAIL] .masonry 1
  assert_equals: \n<div class="masonry" data-expected-width="100" data-expected-height="500">\n      <div class="item" data-expected-width="100" data-expected-height="100">Cell 1</div>\n      <div class="item" data-expected-width="100" data-expected-height="100">Cell 2</div>\n      <div class="item" data-expected-width="100" data-expected-height="100">Cell 3</div>\n      <div class="item" data-expected-width="100" data-expected-height="100">Cell 4</div>\n      <div class="item" data-expected-width="100" data-expected-height="100">Cell 5</div>\n    </div>\nheight expected 100 but got 18
Harness: the test ran to completion.
